+++ ray/tests/common.py
@@ -273,6 +273,9 @@ OPTIONAL_METRICS = [
     'pull_manager.object_request_time.sum',
     'pull_manager.object_request_time.bucket',
     'pull_manager.object_request_time.count',
+    'object_store.size.sum',
+    'object_store.size.count',
+    'object_store.size.bucket',
 ]
 OPTIONAL_METRICS = ['ray.' + m for m in OPTIONAL_METRICS]