+++ ddev/tests/cli/meta/scripts/conftest.py
@@ -5,6 +5,9 @@ import pytest
 
 from ddev.repo.core import Repository
 
+OLD_PYTHON_VERSION = "3.9"
+NEW_PYTHON_VERSION = "3.11"
+
 
 @pytest.fixture
 def fake_repo(tmp_path_factory, config_file, ddev):
@@ -17,31 +20,97 @@ def fake_repo(tmp_path_factory, config_file, ddev):
     write_file(
         repo_path / 'ddev' / 'src' / 'ddev' / 'repo',
         'constants.py',
-        """# (C) Datadog, Inc. 2022-present
+        f"""# (C) Datadog, Inc. 2022-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 CONFIG_DIRECTORY = '.ddev'
 NOT_SHIPPABLE = frozenset(['datadog_checks_dev', 'datadog_checks_tests_helper', 'ddev'])
-FULL_NAMES = {
+FULL_NAMES = {{
     'core': 'integrations-core',
     'extras': 'integrations-extras',
     'marketplace': 'marketplace',
     'agent': 'datadog-agent',
-}
+}}
 
 # This is automatically maintained
-PYTHON_VERSION = '3.9'
+PYTHON_VERSION = '{OLD_PYTHON_VERSION}'
 """,
     )
 
     write_file(
         repo_path / 'dummy',
         'hatch.toml',
-        """[env.collectors.datadog-checks]
+        f"""[env.collectors.datadog-checks]
 
 [[envs.default.matrix]]
-python = ["2.7", "3.9"]
+python = ["2.7", "{OLD_PYTHON_VERSION}"]
 
+""",
+    )
+
+    write_file(
+        repo_path / 'dummy',
+        'pyproject.toml',
+        f"""[project]
+name = "dummy"
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "Intended Audience :: Developers",
+    "Intended Audience :: System Administrators",
+    "License :: OSI Approved :: BSD License",
+    "Natural Language :: English",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 2.7",
+    "Programming Language :: Python :: {OLD_PYTHON_VERSION}",
+]
+""",
+    )
+
+    write_file(
+        repo_path / '.github' / 'workflows',
+        'build-ddev.yml',
+        f"""name: build ddev
+env:
+  APP_NAME: ddev
+  PYTHON_VERSION: "{OLD_PYTHON_VERSION}"
+  PYOXIDIZER_VERSION: "0.24.0"
+""",
+    )
+
+    write_file(
+        repo_path / 'ddev',
+        'pyproject.toml',
+        f"""[tool.black]
+target-version = ["py{OLD_PYTHON_VERSION.replace('.', '')}"]
+
+[tool.ruff]
+target-version = "py{OLD_PYTHON_VERSION.replace('.', '')}"
+""",
+    )
+
+    write_file(
+        repo_path
+        / 'datadog_checks_dev'
+        / 'datadog_checks'
+        / 'dev'
+        / 'tooling'
+        / 'templates'
+        / 'integration'
+        / 'check'
+        / '{check_name}',
+        'pyproject.toml',
+        f"""[project]
+name = "dummy"
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "Intended Audience :: Developers",
+    "Intended Audience :: System Administrators",
+    "License :: OSI Approved :: BSD License",
+    "Natural Language :: English",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 2.7",
+    "Programming Language :: Python :: {OLD_PYTHON_VERSION}",
+]
 """,
     )