+++ ddev/CHANGELOG.md
@@ -9,6 +9,7 @@
 ***Fixed***:
 
 * Fix `ddev env test` so that tests run for all environments properly when no environment is specified ([#16054](https://github.com/DataDog/integrations-core/pull/16054))
+* Include ddev's source code when measuring its coverage ([#16057](https://github.com/DataDog/integrations-core/pull/16057))
 
 ## 5.2.1 / 2023-10-12