+++ ddev/CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fix Github API search query ([#15943](https://github.com/DataDog/integrations-core/pull/15943))
 * Do not modify the Agent build name if provided by the user when running the e2e environments ([#16052](https://github.com/DataDog/integrations-core/pull/16052))
 * Bump the Python version in the dependency provider when bumping the Python version ([#16070](https://github.com/DataDog/integrations-core/pull/16070))
+* Fix `ddev env start` to allow the use of stable releases ([#16077](https://github.com/DataDog/integrations-core/pull/16077))
 
 ## 5.2.1 / 2023-10-12